Function frame_allocator::allocate_frames_by_bytes_at

pub fn allocate_frames_by_bytes_at(
    paddr: PhysicalAddress, 
    num_bytes: usize
) -> Result<AllocatedFrames, &'static str>

Allocates frames starting at the given PhysicalAddress with a size given in number of bytes.

This function still allocates whole frames by rounding up the number of bytes. See allocate_frames_deferred() for more details.