Struct frame_allocator::AllocatedFrame

pub struct AllocatedFrame<'f, P: PageSize> { /* private fields */ }

A reference to a single frame within a range of AllocatedFrames.

The lifetime of this type is tied to the lifetime of its owning AllocatedFrames.