Struct frame_allocator::Frames

pub struct Frames<const S: MemoryState, P: PageSize = Page4K> { /* private fields */ }

A range of contiguous frames in physical memory.

Each Frames object is globally unique, meaning that the owner of a Frames object has globally-exclusive access to the range of frames it contains.

A Frames object can be in one of four states:

Free: frames are owned by the frame allocator and have not been allocated for any use.

Allocated: frames have been removed from the allocator’s free list and are owned elsewhere; they can now be used for mapping purposes.

Mapped: frames have been (and are currently) mapped by a range of virtual memory pages.

Unmapped: frames have been unmapped and can be returned to the frame allocator.

The drop behavior for a Frames object is based on its state:

Free: the frames will be added back to the frame allocator’s free list.

Allocated: the frames will be transitioned into the Free state.

Unmapped: the frames will be transitioned into the Allocated state.

Mapped: currently, Theseus does not actually drop mapped Frames, but rather they are forgotten when they are mapped by virtual pages, and then re-created in the Unmapped state after being unmapped from the page tables.

As such, one can visualize the Frames state diagram as such:

(Free) <---> (Allocated) --> (Mapped) --> (Unmapped) --> (Allocated) <---> (Free)

Ordering and Equality

Frames implements the Ord trait, and its total ordering is ONLY based on its starting Frame. This is useful so we can store Frames in a sorted collection.

Similarly, Frames implements equality traits, Eq and PartialEq, both of which are also based ONLY on the starting Frame of the Frames. Thus, comparing two Frames with the == or != operators may not work as expected. since it ignores their actual range of frames.

Similarly, Frames implements the Borrow trait to return a Frame, not a FrameRange. This is required so we can search for Frames in a sorted collection using a Frame value. It differs from the behavior of the Deref trait which returns a FrameRange.

Implementations

impl Frames<{MemoryState::Free}, Page4K>

pub fn into_allocated_frames(self) -> AllocatedFrames<Page4K>

Consumes this Frames in the Free state and converts them into the Allocated state.

impl<P: PageSize> Frames<{MemoryState::Allocated}, P>

pub fn into_mapped_frames(self) -> MappedFrames<P>

Consumes this Frames in the Allocated state and converts them into the Mapped state. This should only be called once a MappedPages has been created from the Frames.

pub fn as_allocated_frame(&self) -> AllocatedFrame<'_, P>

Returns an AllocatedFrame if this AllocatedFrames object contains only one frame.

Panic

Panics if this AllocatedFrame contains multiple frames or zero frames.

impl Frames<{MemoryState::Unmapped}, Page4K>

pub fn into_allocated_frames(self) -> AllocatedFrames

Consumes this Frames in the Unmapped state and converts them into the Allocated state.

impl<const S: MemoryState, P: PageSize> Frames<S, P>

pub const fn empty() -> Frames<S, P>

Returns a new Frames with an empty range of frames. Can be used as a placeholder, but will not permit any real usage.

pub fn merge(&mut self, other: Self) -> Result<(), Self>

Merges the given other Frames object into this Frames object (self).

This function performs no allocation or re-mapping, it exists for convenience and usability purposes.

The given other must be physically contiguous with self, i.e., come immediately before or after self. That is, either self.start == other.end + 1 or self.end + 1 == other.start must be true.

If either of those conditions are met, self is modified and Ok(()) is returned, otherwise Err(other) is returned.

pub fn split_range( self, frames_to_extract: FrameRange<P> ) -> Result<SplitFrames<S, P>, Self>

Splits up the given Frames into multiple smaller Frames.

Returns a SplitFrames instance containing three Frames:

The range of frames in self that are before the beginning of frames_to_extract.

The Frames containing the requested range of frames, frames_to_extract.

The range of frames in self that are after the end of frames_to_extract.

If frames_to_extract is not contained within self, then self is returned unchanged within an Err.

pub fn split_at(self, at_frame: Frame<P>) -> Result<(Self, Self), Self>

Splits this Frames into two separate Frames objects:

[beginning : at_frame - 1]

[at_frame : end]

This function follows the behavior of core::slice::split_at(), thus, either one of the returned Frames objects may be empty.

If at_frame == self.start, the first returned Frames object will be empty.

If at_frame == self.end + 1, the second returned Frames object will be empty.

Returns an Err containing this Frames if at_frame is otherwise out of bounds, or if self was empty.