Struct frame_allocator::PhysicalMemoryRegion

pub struct PhysicalMemoryRegion {
    pub frames: FrameRange,
    pub typ: MemoryRegionType,
}

PhysicalMemoryRegion represents a range of contiguous frames in physical memory for bookkeeping purposes. It does not give access to the underlying frames.

Ordering and Equality

PhysicalMemoryRegion implements the Ord trait, and its total ordering is ONLY based on its starting Frame. This is useful so we can store PhysicalMemoryRegions in a sorted collection.

Similarly, PhysicalMemoryRegion implements equality traits, Eq and PartialEq, both of which are also based ONLY on the starting Frame of the PhysicalMemoryRegion. Thus, comparing two PhysicalMemoryRegions with the == or != operators may not work as expected. since it ignores their actual range of frames.

Fields
frames: FrameRange

The Frames covered by this region, an inclusive range.

typ: MemoryRegionType

The type of this memory region, e.g., whether it’s in a free or reserved region.

Implementations

impl PhysicalMemoryRegion

pub fn new(frames: FrameRange, typ: MemoryRegionType) -> PhysicalMemoryRegion