Type Definition frame_allocator::AllocatedFrames

pub type AllocatedFrames = Frames<{ MemoryState::Allocated }>;

A type alias for Frames in the Allocated state.

Implementations

impl AllocatedFrames

pub fn into_mapped_frames(self) -> MappedFrames

Consumes this Frames in the Allocated state and converts them into the Mapped state. This should only be called once a MappedPages has been created from the Frames.

pub fn as_allocated_frame(&self) -> AllocatedFrame<'_>

Returns an AllocatedFrame if this AllocatedFrames object contains only one frame.

Panic

Panics if this AllocatedFrame contains multiple frames or zero frames.