Type Alias frame_allocator::UnmappedFrames

pub type UnmappedFrames<P: PageSize = Page4K> = Frames<{ MemoryState::Unmapped }, P>;

A type alias for Frames in the Unmapped state.

Aliased Type

struct UnmappedFrames<P: PageSize = Page4K> { /* private fields */ }

Implementations

impl UnmappedFrames

pub fn into_allocated_frames(self) -> AllocatedFrames

Consumes this Frames in the Unmapped state and converts them into the Allocated state.

impl Frames<{MemoryState::Free}, Page4K>

pub fn into_allocated_frames(self) -> AllocatedFrames<Page4K>

Consumes this Frames in the Free state and converts them into the Allocated state.

impl<P: PageSize> Frames<{MemoryState::Allocated}, P>

pub fn into_mapped_frames(self) -> MappedFrames<P>

Consumes this Frames in the Allocated state and converts them into the Mapped state. This should only be called once a MappedPages has been created from the Frames.

pub fn as_allocated_frame(&self) -> AllocatedFrame<'_, P>

Returns an AllocatedFrame if this AllocatedFrames object contains only one frame.

Panic

Panics if this AllocatedFrame contains multiple frames or zero frames.

impl<const S: MemoryState, P: PageSize> Frames<S, P>

pub const fn empty() -> Frames<S, P>

Returns a new Frames with an empty range of frames. Can be used as a placeholder, but will not permit any real usage.

pub fn merge(&mut self, other: Self) -> Result<(), Self>

Merges the given other Frames object into this Frames object (self).

This function performs no allocation or re-mapping, it exists for convenience and usability purposes.

The given other must be physically contiguous with self, i.e., come immediately before or after self. That is, either self.start == other.end + 1 or self.end + 1 == other.start must be true.

If either of those conditions are met, self is modified and Ok(()) is returned, otherwise Err(other) is returned.

pub fn split_range( self, frames_to_extract: FrameRange<P> ) -> Result<SplitFrames<S, P>, Self>

Splits up the given Frames into multiple smaller Frames.

Returns a SplitFrames instance containing three Frames:

The range of frames in self that are before the beginning of frames_to_extract.

The Frames containing the requested range of frames, frames_to_extract.

The range of frames in self that are after the end of frames_to_extract.

If frames_to_extract is not contained within self, then self is returned unchanged within an Err.

pub fn split_at(self, at_frame: Frame<P>) -> Result<(Self, Self), Self>

Splits this Frames into two separate Frames objects:

[beginning : at_frame - 1]

[at_frame : end]

This function follows the behavior of core::slice::split_at(), thus, either one of the returned Frames objects may be empty.

If at_frame == self.start, the first returned Frames object will be empty.

If at_frame == self.end + 1, the second returned Frames object will be empty.

Returns an Err containing this Frames if at_frame is otherwise out of bounds, or if self was empty.