Crate frame_allocator

Provides an allocator for physical memory frames. The minimum unit of allocation is a single frame.

This is currently a modified and more complex version of the page_allocator crate. TODO: extract the common code and create a generic allocator that can be specialized to allocate pages or frames.

This also supports early allocation of frames before heap allocation is available, and does so behind the scenes using the same single interface. Early pre-heap allocations are limited to tracking a small number of available chunks (currently 32).

Once heap allocation is available, it uses a dynamically-allocated list of frame chunks to track allocations.

The core allocation function is allocate_frames_deferred(), but there are several convenience functions that offer simpler interfaces for general usage.

Notes and Missing Features

This allocator currently does not merge freed chunks (de-fragmentation). We don’t need to do so until we actually run out of address space or until a requested address is in a chunk that needs to be merged.

Structs

AllocatedFrame

A reference to a single frame within a range of AllocatedFrames.

AllocatedFrames

Represents a range of allocated physical memory [Frame]s; derefs to [FrameRange].

AllocatedFramesIter

An iterator over each AllocatedFrame in a range of AllocatedFrames.

DeferredAllocAction

A series of pending actions related to frame allocator bookkeeping, which may result in heap allocation.

PhysicalMemoryRegion

A region of physical memory.

Enums

MemoryRegionType

Types of physical memory. See each variant’s documentation.

Functions

allocate_frames

Allocates the given number of frames with no constraints on the starting physical address.

allocate_frames_at

Allocates the given number of frames starting at (inclusive of) the frame containing the given PhysicalAddress.

allocate_frames_by_bytes

Allocates frames with no constraints on the starting physical address, with a size given by the number of bytes.

allocate_frames_by_bytes_at

Allocates frames starting at the given PhysicalAddress with a size given in number of bytes.

allocate_frames_by_bytes_deferred

Similar to allocated_frames_deferred(), but accepts a size value for the allocated frames in number of bytes instead of number of frames.

allocate_frames_deferred

The core frame allocation routine that allocates the given number of physical frames, optionally at the requested starting PhysicalAddress.

init

Initialize the frame allocator with the given list of available and reserved physical memory regions.